azure-function-app/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob
azure-core
azure-cosmos

azure-common
azure-communication-email
azure-mgmt-core

pandas
openpyxl
pyodbc
sqlalchemy
pdfservices-sdk

